requirements.txt
streamlit
pymongo
python-dotenv
langchain
langchain-google-genai
google-generativeai